Add files via upload
The “freemium” business model — widely used by online services such as LinkedIn, Match.com, Dropbox, and music-listening sites — divides user populations into groups that use the service for free and groups that pay a fee for additional features. Key points related to the freemium model: 
-Free accounts are monetized using online advertising
-Premium subscribers (those paying a fee) are typically 24 times more profitable than free users
However, premium subscribers are rare.

Given the higher profitability of premium subscribers, it is generally in the interest of company to motivate users to go from “free to fee”; that is, convert free accounts to premium subscribers. The purpose in regards to this case is to analyze the data for potential insight to inform a “free-to-fee” strategy.